Crate diem_data_client

Modules

diemnet

Structs

AdvertisedData

A summary of all data that is currently advertised in the network.

GlobalDataSummary

A snapshot of the global state of data available in the Diem network.

OptimalChunkSizes

Holds the optimal chunk sizes that clients should use when requesting data. This makes the request more likely to succeed.

Response

A response from the Data Client for a single API call.

ResponseContext

Enums

Error

An error returned by the Diem Data Client for failed API calls.

ResponseError

A response error that users of the Diem Data Client can use to notify the Data Client about invalid or malformed responses.

ResponsePayload

The different data client response payloads as an enum.

Traits

DiemDataClient

The API offered by the Diem Data Client.

ResponseCallback

A callback that lets the consumer provide error feedback about a response. Typically, this will contain a reference to the underlying data client and any additional request context needed to update internal scoring.

Type Definitions

ResponseId

Result