Trait diem_data_client::DiemDataClient

pub trait DiemDataClient {
    #[must_use]
    fn get_account_states_with_proof<'life0, 'async_trait>(
        &'life0 self, 
        version: u64, 
        start_index: HashValue, 
        end_index: HashValue
    ) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_epoch_ending_ledger_infos<'life0, 'async_trait>(
        &'life0 self, 
        start_epoch: u64, 
        end_epoch: u64
    ) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_global_data_summary(&self) -> Result<DataClientResponse, Error>;

    #[must_use]
    fn get_number_of_account_states<'life0, 'async_trait>(
        &'life0 self, 
        version: u64
    ) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_transaction_outputs_with_proof<'life0, 'async_trait>(
        &'life0 self, 
        proof_version: u64, 
        start_version: u64, 
        end_version: u64
    ) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_transactions_with_proof<'life0, 'async_trait>(
        &'life0 self, 
        proof_version: u64, 
        start_version: u64, 
        end_version: u64, 
        include_events: bool
    ) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn notify_bad_response<'life0, 'async_trait>(
        &'life0 self, 
        response_id: u64, 
        response_error: ResponseError
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

The API offered by the Diem Data Client.

Required methods

#[must_use]
fn get_account_states_with_proof<'life0, 'async_trait>(
&'life0 self,
version: u64,
start_index: HashValue,
end_index: HashValue
) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a single account states chunk with proof, containing the accounts from start to end index (inclusive) at the specified version. The proof version is the same as the specified version.

#[must_use]
fn get_epoch_ending_ledger_infos<'life0, 'async_trait>(
&'life0 self,
start_epoch: u64,
end_epoch: u64
) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns all epoch ending ledger infos between start and end (inclusive). If the data cannot be fetched (e.g., the number of epochs is too large), an error is returned.

fn get_global_data_summary(&self) -> Result<DataClientResponse, Error>

Returns a global summary of the data currently available in the network.

#[must_use]
fn get_number_of_account_states<'life0, 'async_trait>(
&'life0 self,
version: u64
) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the number of account states at the specified version.

#[must_use]
fn get_transaction_outputs_with_proof<'life0, 'async_trait>(
&'life0 self,
proof_version: u64,
start_version: u64,
end_version: u64
) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a transaction output list with proof object, with transaction outputs from start to end versions (inclusive). The proof is relative to the specified proof_version. If the data cannot be fetched (e.g., the number of transaction outputs is too large), an error is returned.

#[must_use]
fn get_transactions_with_proof<'life0, 'async_trait>(
&'life0 self,
proof_version: u64,
start_version: u64,
end_version: u64,
include_events: bool
) -> Pin<Box<dyn Future<Output = Result<DataClientResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a transaction list with proof object, with transactions from start to end versions (inclusive). The proof is relative to the specified proof_version. If include_events is true, events are included in the proof. If the data cannot be fetched (e.g., the number of transactions is too large), an error is returned.

#[must_use]
fn notify_bad_response<'life0, 'async_trait>(
&'life0 self,
response_id: u64,
response_error: ResponseError
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Notifies the Diem Data Client about a previously received response that was bad (e.g., invalid or malformed).

Note: this is required because the Diem Data Client can only fetch data from peers in the network, but it is not able to fully verify that the given data responses are valid (e.g., it is unable to verify proofs). This API call provides a simple feedback mechanism for users of the Diem Data Client to alert it to bad responses so that the peers responsible for providing this data can be penalized. The response_id is the handle used by clients to notify the Diem Data Client of invalid responses.